src/holopy/quantum/channels.py
"""
Quantum channel dynamics with holographic decoherence.
"""
from typing import List, Tuple, Dict, Optional
import numpy as np
from scipy.linalg import expm
import logging
from dataclasses import dataclass
from ..config.constants import (
    INFORMATION_GENERATION_RATE,
    PLANCK_CONSTANT,
    SPEED_OF_LIGHT,
    COUPLING_CONSTANT
)

logger = logging.getLogger(__name__)

@dataclass
class ChannelMetrics:
    """Metrics for quantum channel evolution."""
    channel_fidelity: float
    coherence_time: float
    decoherence_rate: float
    information_loss: float
    entropy_production: float

class HolographicChannel:
    """Implements quantum channels with holographic decoherence."""
    
    def __init__(
        self,
        n_qubits: int,
        temperature: float = 0.0,
        coupling_strength: Optional[float] = None
    ):
        """
        Initialize quantum channel.
        
        Args:
            n_qubits: Number of qubits
            temperature: Environment temperature
            coupling_strength: System-environment coupling
        """
        self.n_qubits = n_qubits
        self.temperature = temperature
        self.coupling_strength = coupling_strength or COUPLING_CONSTANT
        
        # Initialize channel parameters
        self._initialize_channel()
        
        logger.info(
            f"Initialized HolographicChannel for {n_qubits} qubits at "
            f"T={temperature:.2f}"
        )
    
    def _initialize_channel(self) -> None:
        """Initialize quantum channel parameters."""
        try:
            # Calculate basic decoherence rates
            self.dephasing_rate = (
                INFORMATION_GENERATION_RATE * 
                (1 + self.temperature / PLANCK_CONSTANT)
            )
            
            self.dissipation_rate = (
                self.coupling_strength * 
                np.tanh(PLANCK_CONSTANT / (2 * self.temperature))
                if self.temperature > 0 else 0
            )
            
            # Initialize Lindblad operators
            self._initialize_lindblad_operators()
            
            logger.debug(
                f"Initialized channel with dephasing rate {self.dephasing_rate:.2e}"
            )
            
        except Exception as e:
            logger.error(f"Channel initialization failed: {str(e)}")
            raise
    
    def evolve_state(
        self,
        state: np.ndarray,
        time: float
    ) -> Tuple[np.ndarray, ChannelMetrics]:
        """
        Evolve quantum state through channel.
        
        Args:
            state: Initial quantum state
            time: Evolution time
            
        Returns:
            Tuple of (evolved_state, channel_metrics)
        """
        try:
            # Calculate evolution operator
            evolution = self._calculate_evolution_operator(time)
            
            # Apply evolution with holographic corrections
            initial_state = state.copy()
            evolved_state = evolution @ state
            
            # Calculate channel metrics
            metrics = self._calculate_channel_metrics(
                initial_state,
                evolved_state,
                time
            )
            
            logger.debug(
                f"Evolved state for time {time:.4f} with "
                f"fidelity {metrics.channel_fidelity:.4f}"
            )
            
            return evolved_state, metrics
            
        except Exception as e:
            logger.error(f"State evolution failed: {str(e)}")
            raise
    
    def _initialize_lindblad_operators(self) -> None:
        """Initialize Lindblad operators for decoherence."""
        try:
            dim = 2**self.n_qubits
            self.lindblad_operators = []
            
            # Single-qubit dephasing operators
            for i in range(self.n_qubits):
                op = np.zeros((dim, dim), dtype=complex)
                for j in range(dim):
                    if j & (1 << i):
                        op[j,j] = 1
                    else:
                        op[j,j] = -1
                self.lindblad_operators.append(
                    np.sqrt(self.dephasing_rate/2) * op
                )
            
            # Dissipation operators if temperature > 0
            if self.temperature > 0:
                for i in range(self.n_qubits):
                    # Lowering operator
                    op = np.zeros((dim, dim), dtype=complex)
                    for j in range(dim):
                        if j & (1 << i):
                            k = j & ~(1 << i)
                            op[k,j] = 1
                    self.lindblad_operators.append(
                        np.sqrt(self.dissipation_rate) * op
                    )
                    
                    # Raising operator
                    op = np.zeros((dim, dim), dtype=complex)
                    for j in range(dim):
                        if not (j & (1 << i)):
                            k = j | (1 << i)
                            op[k,j] = 1
                    self.lindblad_operators.append(
                        np.sqrt(self.dissipation_rate * 
                               np.exp(-PLANCK_CONSTANT/self.temperature)) * op
                    )
            
            logger.debug(f"Initialized {len(self.lindblad_operators)} Lindblad operators")
            
        except Exception as e:
            logger.error(f"Lindblad operator initialization failed: {str(e)}")
            raise
    
    def _calculate_evolution_operator(self, time: float) -> np.ndarray:
        """Calculate quantum channel evolution operator."""
        try:
            dim = 2**self.n_qubits
            
            # Construct Liouvillian
            liouvillian = np.zeros((dim*dim, dim*dim), dtype=complex)
            
            # Add Hamiltonian evolution
            hamiltonian = self._system_hamiltonian()
            h_term = -1j/PLANCK_CONSTANT * (
                np.kron(hamiltonian, np.eye(dim)) - 
                np.kron(np.eye(dim), hamiltonian.conj())
            )
            liouvillian += h_term
            
            # Add dissipative terms
            for L in self.lindblad_operators:
                l_term = (
                    np.kron(L, L.conj()) -
                    0.5 * np.kron(L.conj().T @ L, np.eye(dim)) -
                    0.5 * np.kron(np.eye(dim), L.T @ L.conj())
                )
                liouvillian += l_term
            
            # Calculate evolution
            evolution = expm(liouvillian * time)
            
            # Reshape to superoperator form
            evolution = evolution.reshape(dim, dim, dim, dim)
            evolution = np.transpose(evolution, (0, 2, 1, 3))
            evolution = evolution.reshape(dim*dim, dim*dim)
            
            return evolution
            
        except Exception as e:
            logger.error(f"Evolution operator calculation failed: {str(e)}")
            raise
    
    def _system_hamiltonian(self) -> np.ndarray:
        """Calculate system Hamiltonian with holographic corrections."""
        try:
            dim = 2**self.n_qubits
            hamiltonian = np.zeros((dim, dim), dtype=complex)
            
            # Add local terms
            for i in range(self.n_qubits):
                h_local = np.zeros((dim, dim), dtype=complex)
                for j in range(dim):
                    if j & (1 << i):
                        h_local[j,j] = 1
                hamiltonian += h_local
            
            # Add holographic corrections
            hamiltonian *= np.exp(-INFORMATION_GENERATION_RATE * 
                                np.arange(dim)/dim)[:, None]
            
            return hamiltonian
            
        except Exception as e:
            logger.error(f"Hamiltonian calculation failed: {str(e)}")
            raise 